renaming bigCard, renaming course_card.dart, moving the cards to the right folder
