+++ lib/views/course_view/components/course_card.dart
