+++ lib/views/course_view/course_detail_view/stream_card.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:gocast_mobile/base/networking/api/gocast/api_v2.pb.dart';
-import 'package:gocast_mobile/views/course_view/components/big_card.dart';
+import 'package:gocast_mobile/views/course_view/components/baseCard.dart';
 
-class StreamCard extends BigCard {
+class StreamCard extends BaseCard {
   final Stream stream;
 
   const StreamCard({
